+++ test/test_rs485/test_matching_bytes.h
@@ -26,7 +26,7 @@ TEST_F(ProtocolMatchingBytesTest, single_byte_odd) {
   buffer << 0x01;
   bus.fetch();
 
-  int endIndex = 0;
+  size_t endIndex = 0;
   EXPECT_EQ(protocol.isPacket(bus, 0, endIndex), PacketStatus::NO);
 }
 
@@ -34,7 +34,7 @@ TEST_F(ProtocolMatchingBytesTest, single_byte_even) {
   buffer << 0x02;
   bus.fetch();
 
-  int endIndex = 0;
+  size_t endIndex = 0;
   EXPECT_EQ(protocol.isPacket(bus, 0, endIndex), PacketStatus::NOT_ENOUGH_BYTES);
 }
 
@@ -42,7 +42,7 @@ TEST_F(ProtocolMatchingBytesTest, various_packets) {
   buffer << 0x01 << 0x02 << 0x03 << 0x01 << 0xff << 0xfe << 0x02;
   bus.fetch();
 
-  int endIndex = 6;
+  size_t endIndex = 6;
   EXPECT_EQ(protocol.isPacket(bus, 0, endIndex), PacketStatus::YES);
   EXPECT_EQ(endIndex, 3);
